+++ src/itemAction/index.unit.tsx
@@ -43,7 +43,7 @@ describe('ItemAction', () => {
   it('Should set tag to button by default', () => {
     const wrapper = shallow(<ItemAction {...defaultProps} />)
 
-    expect(wrapper.prop('tag')).toEqual(<button />)
+    expect(wrapper.prop('tag')).toEqual(<button type="button"/>)
   })
 
   it('Should forward the tag to item', () => {